be/requirements.txt
fastapi
uvicorn
requests
python-dotenv
fredapi
pandas
re
bs4